// back/src/models/member.ts
import { BaseModel } from "./base";

export class Member extends BaseModel {

    first_name: string;
    middle_name: string;
    last_name: string;
    occupation: string;
    email: string;
    mobile_no: string;
    location: string;
    password: string;
    usertype: string;
    group_id: number;
    group_category_ids: any;

    static tableName = 'members';

}